+++ src/graphics/mesh.rs
@@ -68,7 +68,7 @@ unsafe impl Zeroable for Vertex {}
 /// The expected usage of a GPU buffer.
 ///
 /// The GPU may optionally use this to optimize data storage and access.
-#[derive(Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum BufferUsage {
     /// The buffer's data is not expected to change after creation.
     Static,
@@ -81,7 +81,7 @@ pub enum BufferUsage {
 }
 
 /// The ordering of the vertices in a piece of geometry.
-#[derive(Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum VertexWinding {
     /// The vertices are in clockwise order.
     Clockwise,