+++ CHANGELOG.md
@@ -6,7 +6,7 @@ From 0.4.0 onwards, all breaking changes will be explicitly labelled, to make it
 
 This project adheres to Semantic Versioning.
 
-## [Upcoming]
+## [0.8.0] - 2023-03-17
 
 ### Added
 
@@ -908,7 +908,8 @@ for. This can be useful when implementing more complex animation behaviors. ([@V
 
 * Initial release!
 
-[Upcoming]: https://github.com/17cupsofcoffee/tetra/compare/0.7.0..HEAD
+[Upcoming]: https://github.com/17cupsofcoffee/tetra/compare/0.8.0..HEAD
+[0.7.0]: https://github.com/17cupsofcoffee/tetra/compare/0.7.0..0.8.0
 [0.7.0]: https://github.com/17cupsofcoffee/tetra/compare/0.6.7..0.7.0
 [0.6.7]: https://github.com/17cupsofcoffee/tetra/compare/0.6.6..0.6.7
 [0.6.6]: https://github.com/17cupsofcoffee/tetra/compare/0.6.5..0.6.6